refactor: extract application layer helpers
